Fix eq instance and properly test it
